+++ python/pydantable/__init__.py
@@ -29,4 +29,4 @@ __all__ = [
     "read_ipc",
     "read_parquet",
 ]
-__version__ = "0.19.0"
+__version__ = "0.20.0"